Added data curation, began data EDA
Okay so the main thing is that it turns out that the library gtrendsR absolutely blows chunks. It works very well, for a few queries. Then it breaks. Its also hella slow. Reinstalling it seems to help. With all that in mind, I created a function called "add_term" that pulls the google trend data using the above library and automatically adds a column to the data frame called whatever the term is plus "_hits" appended to it. Note that this only shows data for the past 90 days, so the January ones will all be NA.  add_term should be called as infrequently as possibly, to avoid the buggy library we're using. Other than that, I threw up a few graphs and stuff, we should probably add other ones that are more interesting but the ones I put up show a general format of how to use the dataframe to create plots. In other news, some of my code is buggy. For some reason, Alabama shows up like 4000 extra times for no reason in the table, or at least a reason beyond my current sleep deprived mind's comprehension. That's all